# wasi: optimizes args/environ parsing