While most compilers will only read args/environ once, tools like WAGI make heavy use of environment, possibly dozens of long variables. This optimizes both args and environ for this reason and also to setup for optimizing other functions.

Here are the notable changes:
* eagerly coerce to byte slices instead of strings
* re-use null terminated length for writing values
* avoid loops that call mem.WriteXXX internally

Before:
```
Benchmark_ArgsEnviron
Benchmark_ArgsEnviron/args_get
Benchmark_ArgsEnviron/args_get-16         	 6147727	       187.9 ns/op	      72 B/op	       5 allocs/op
Benchmark_ArgsEnviron/args_sizes_get
Benchmark_ArgsEnviron/args_sizes_get-16   	11869351	       100.7 ns/op	      24 B/op	       2 allocs/op
Benchmark_ArgsEnviron/environ_get
Benchmark_ArgsEnviron/environ_get-16      	 6202508	       191.0 ns/op	      72 B/op	       5 allocs/op
Benchmark_ArgsEnviron/environ_sizes_get
Benchmark_ArgsEnviron/environ_sizes_get-16         	11792656	        99.71 ns/op	      24 B/op	       2 allocs/op
```

After:
```
Benchmark_ArgsEnviron
Benchmark_ArgsEnviron/args_get
Benchmark_ArgsEnviron/args_get-16         	 9542246	       119.2 ns/op	      24 B/op	       2 allocs/op
Benchmark_ArgsEnviron/args_sizes_get
Benchmark_ArgsEnviron/args_sizes_get-16   	11963056	       105.7 ns/op	      24 B/op	       2 allocs/op
Benchmark_ArgsEnviron/environ_get
Benchmark_ArgsEnviron/environ_get-16      	 9742885	       123.0 ns/op	      24 B/op	       2 allocs/op
Benchmark_ArgsEnviron/environ_sizes_get
Benchmark_ArgsEnviron/environ_sizes_get-16         	12100869	        97.06 ns/op	      24 B/op	       2 allocs/op
```