@codefromthecrypt 2022-12-05
ps for those interested in inlining, there was discussion but closed without doing anything on go:inline https://github.com/golang/go/issues/21536 If you build with `-gcflags="-m"` you can see things including what was inlined.